+++ DelvCD/Config/CooldownTrigger.cs
@@ -1,4 +1,5 @@
 using Dalamud.Game.ClientState;
+using Dalamud.Interface;
 using DelvCD.Helpers;
 using ImGuiNET;
 using System;
@@ -11,6 +12,8 @@ namespace DelvCD.Config
 {
     public class CooldownTrigger : TriggerOptions
     {
+        [JsonIgnore] private float _scale => ImGuiHelpers.GlobalScale;
+
         [JsonIgnore] private static readonly string[] _comboOptions = new[] { "Ready", "Not Ready" };
         [JsonIgnore] private static readonly string[] _usableOptions = new[] { "Usable", "Not Usable" };
         [JsonIgnore] private static readonly string[] _rangeOptions = new[] { "In Range", "Not in Range" };
@@ -164,9 +167,9 @@ namespace DelvCD.Config
             DrawHelpers.DrawSpacing(1);
             ImGui.Text("Trigger Conditions");
             string[] operatorOptions = TriggerOptions.OperatorOptions;
-            float optionsWidth = 100 + padX;
-            float opComboWidth = 55;
-            float valueInputWidth = 45;
+            float optionsWidth = 100 * _scale + padX;
+            float opComboWidth = 55 * _scale;
+            float valueInputWidth = 45 * _scale;
             float padWidth = 0;
 
             DrawHelpers.DrawNestIndicator(1);